Added stale data handling
When a `stale_ttl` option is passed either to the LRU constructor, or
as a `:get()` option, we might now return stale data whenever the L3
callback fails.

This is signaled to the called by adding a 4th "layer" (the stale layer).
The rationale is that in practice, `nil, err` returns for functions are
for unusual but not exceptional circumstances (e.g. timeout) and the `get`
caller should not be disrupted by that when this happens.